+++ console_backend/src/utils/formatters.rs
